+++ Snowflake.Data.Tests/IntegrationTests/SFConnectionPoolT.cs
@@ -2,7 +2,7 @@
  * Copyright (c) 2012-2021 Snowflake Computing Inc. All rights reserved.
  */
 
-namespace Snowflake.Data.Tests
+namespace Snowflake.Data.Tests.IntegrationTests
 {
     using NUnit.Framework;
     using Snowflake.Data.Client;
