+++ Snowflake.Data.Tests/IntegrationTests/SFDbTransactionIT.cs
@@ -2,7 +2,7 @@
  * Copyright (c) 2012-2019 Snowflake Computing Inc. All rights reserved.
  */
 
-namespace Snowflake.Data.Tests
+namespace Snowflake.Data.Tests.IntegrationTests
 {
     using System.Data;
     using System.Data.Common;
